+++ src/metanetx_sdk/extract.py
@@ -25,23 +25,23 @@ import pandas as pd
 from . import data
 
 
-def extract_chemical_registry_mapping():
-    """Return the packaged chemical registry mapping."""
-    with open_text(data, "chem_registry.tsv") as handle:
+def extract_chemical_prefix_mapping():
+    """Return the packaged chemical prefix mapping."""
+    with open_text(data, "chem_prefix.tsv") as handle:
         mapping = pd.read_csv(handle, sep="\t", index_col="mnx", squeeze=True)
     return mapping
 
 
-def extract_compartment_registry_mapping():
-    """Return the packaged compartment registry mapping."""
-    with open_text(data, "comp_registry.tsv") as handle:
+def extract_compartment_prefix_mapping():
+    """Return the packaged compartment prefix mapping."""
+    with open_text(data, "comp_prefix.tsv") as handle:
         mapping = pd.read_csv(handle, sep="\t", index_col="mnx", squeeze=True)
     return mapping
 
 
-def extract_reaction_registry_mapping():
-    """Return the packaged reaction registry mapping."""
-    with open_text(data, "reac_registry.tsv") as handle:
+def extract_reaction_prefix_mapping():
+    """Return the packaged reaction prefix mapping."""
+    with open_text(data, "reac_prefix.tsv") as handle:
         mapping = pd.read_csv(handle, sep="\t", index_col="mnx", squeeze=True)
     return mapping