+++ src/metanetx_sdk/transform/compartment.py
@@ -25,69 +25,69 @@ import pandas as pd
 logger = logging.getLogger(__name__)
 
 
-def transform_cell_cycle_ontology_registry(table: pd.DataFrame):
+def transform_cell_cycle_ontology_prefix(table: pd.DataFrame):
     """Transform all CCO terms."""
-    mask = table["registry"] == "cco"
-    table.loc[mask, "accession"] = "CCO:" + table.loc[mask, "accession"]
+    mask = table["prefix"] == "cco"
+    table.loc[mask, "identifier"] = "CCO:" + table.loc[mask, "identifier"]
 
 
-def transform_go_registry(table: pd.DataFrame):
+def transform_go_prefix(table: pd.DataFrame):
     """Transform all GO terms."""
-    mask = table["registry"] == "go"
-    table.loc[mask, "accession"] = "GO:" + table.loc[mask, "accession"]
+    mask = table["prefix"] == "go"
+    table.loc[mask, "identifier"] = "GO:" + table.loc[mask, "identifier"]
 
 
-def transform_metanetx_registry(table: pd.DataFrame):
+def transform_metanetx_prefix(table: pd.DataFrame):
     """Transform all MetaNetX identifiers."""
-    # MetaNetX identifiers themselves have no registry. So we add it.
-    mnx_mask = table["accession"].isnull()
-    table.loc[mnx_mask, "accession"] = table.loc[mnx_mask, "registry"]
-    table.loc[mnx_mask, "registry"] = "metanetx.compartment"
+    # MetaNetX identifiers themselves have no prefix. So we add it.
+    mnx_mask = table["identifier"].isnull()
+    table.loc[mnx_mask, "identifier"] = table.loc[mnx_mask, "prefix"]
+    table.loc[mnx_mask, "prefix"] = "metanetx.compartment"
 
 
 def transform_compartment_properties(
-    compartments: pd.DataFrame, registry_mapping: Mapping
+    compartments: pd.DataFrame, prefix_mapping: Mapping
 ) -> pd.DataFrame:
     """Transform the MetaNetX compartment properties."""
     df = compartments.copy()
     # Cross references have a prefix.
     # We split the prefixes so that we know the actual data sources.
-    df[["registry", "accession"]] = df["source"].str.split(":", n=1, expand=True)
+    df[["prefix", "identifier"]] = df["source"].str.split(":", n=1, expand=True)
     # Map all source databases to MIRIAM compliant versions.
-    for registry in df.loc[df["accession"].notnull(), "registry"].unique():
-        if registry in registry_mapping:
-            df.loc[df["registry"] == registry, "registry"] = registry_mapping[registry]
+    for prefix in df.loc[df["identifier"].notnull(), "prefix"].unique():
+        if prefix in prefix_mapping:
+            df.loc[df["prefix"] == prefix, "prefix"] = prefix_mapping[prefix]
         else:
             logger.warning(
-                "The resource prefix '%s' does not appear in the mapping.", registry
+                "The resource prefix '%s' does not appear in the mapping.", prefix
             )
-    transform_cell_cycle_ontology_registry(df)
-    transform_go_registry(df)
-    transform_metanetx_registry(df)
+    transform_cell_cycle_ontology_prefix(df)
+    transform_go_prefix(df)
+    transform_metanetx_prefix(df)
     del df["source"]
     logger.debug(df.head())
     return df
 
 
 def transform_compartment_cross_references(
-    references: pd.DataFrame, registry_mapping: Mapping
+    references: pd.DataFrame, prefix_mapping: Mapping
 ) -> pd.DataFrame:
     """Transform the MetaNetX compartment cross-references."""
     df = references.copy()
     # Cross references have a prefix.
     # We split the prefixes so that we know the actual data sources.
-    df[["registry", "accession"]] = df["xref"].str.split(":", n=1, expand=True)
+    df[["prefix", "identifier"]] = df["xref"].str.split(":", n=1, expand=True)
     # Map all xref databases to MIRIAM compliant versions.
-    for registry in df.loc[df["accession"].notnull(), "registry"].unique():
-        if registry in registry_mapping:
-            df.loc[df["registry"] == registry, "registry"] = registry_mapping[registry]
+    for prefix in df.loc[df["identifier"].notnull(), "prefix"].unique():
+        if prefix in prefix_mapping:
+            df.loc[df["prefix"] == prefix, "prefix"] = prefix_mapping[prefix]
         else:
             logger.warning(
-                "The resource prefix '%s' does not appear in the mapping.", registry
+                "The resource prefix '%s' does not appear in the mapping.", prefix
             )
-    transform_cell_cycle_ontology_registry(df)
-    transform_go_registry(df)
-    transform_metanetx_registry(df)
+    transform_cell_cycle_ontology_prefix(df)
+    transform_go_prefix(df)
+    transform_metanetx_prefix(df)
     del df["xref"]
     logger.debug(df.head())
     return df